products/requirements.txt
flask==2.2.5
psycopg2-binary==2.9.7
gunicorn==20.1.0
flask-cors==3.0.10
